[Scripts] Toybox Updates

+++ unpublishedScripts/hiddenEntityReset.js
@@ -319,6 +319,9 @@
             userData: JSON.stringify({
               resetMe: {
                 resetMe: true
+              },
+              grabbableKey: {
+                invertSolidWhileHeld: true
               }
             })
           });
@@ -378,7 +381,8 @@
       var MODEL_URL = 'http://hifi-public.s3.amazonaws.com/models/ping_pong_gun/target.fbx';
       var COLLISION_HULL_URL = 'http://hifi-public.s3.amazonaws.com/models/ping_pong_gun/target_collision_hull.obj';
 
-      var RESET_DISTANCE = 1;
+      var MINIMUM_MOVE_LENGTH = 0.05;
+      var RESET_DISTANCE = 0.5;
       var TARGET_USER_DATA_KEY = 'hifi-ping_pong_target';
       var NUMBER_OF_TARGETS = 6;
       var TARGETS_PER_ROW = 3;
@@ -391,7 +395,6 @@
 
       var VERTICAL_SPACING = TARGET_DIMENSIONS.y + 0.5;
       var HORIZONTAL_SPACING = TARGET_DIMENSIONS.z + 0.5;
-
 
       var startPosition = {
         x: 548.68,
@@ -413,6 +416,9 @@
         userData: JSON.stringify({
           resetMe: {
             resetMe: true
+          },
+          grabbableKey: {
+            grabbable: false
           }
         })
       });
@@ -420,6 +426,8 @@
       var targets = [];
 
       var originalPositions = [];
+
+      var lastPositions = [];
 
       function addTargets() {
         var i;
@@ -437,6 +445,7 @@
           position.y = startPosition.y - (row * VERTICAL_SPACING);
 
           originalPositions.push(position);
+          lastPositions.push(position);
 
           var targetProperties = {
             name: 'Target',
@@ -452,6 +461,9 @@
             userData: JSON.stringify({
               resetMe: {
                 resetMe: true
+              },
+              grabbableKey: {
+                grabbable: false
               }
             })
           };
@@ -468,7 +480,11 @@
           var distance = Vec3.subtract(originalPosition, currentPosition);
           var length = Vec3.length(distance);
 
-          if (length > RESET_DISTANCE) {
+          var moving = Vec3.length(Vec3.subtract(currentPosition, lastPositions[index]));
+
+          lastPositions[index] = currentPosition;
+
+          if (length > RESET_DISTANCE && moving < MINIMUM_MOVE_LENGTH) {
 
             Entities.deleteEntity(target);
 
@@ -486,11 +502,14 @@
               userData: JSON.stringify({
                 resetMe: {
                   resetMe: true
+                },
+                grabbableKey: {
+                  grabbable: false
                 }
               })
             };
-            var target = Entities.addEntity(targetProperties);
-            targets[index] = target;
+
+            targets[index] = Entities.addEntity(targetProperties);
 
           }
         });
@@ -548,6 +567,9 @@
         userData: JSON.stringify({
           resetMe: {
             resetMe: true
+          },
+          grabbableKey: {
+            grabbable: false
           }
         })
       });
@@ -583,7 +605,11 @@
         userData: JSON.stringify({
           resetMe: {
             resetMe: true
+          },
+          grabbableKey: {
+            invertSolidWhileHeld: true
           }
+
         })
       });
 
@@ -851,6 +877,9 @@
         userData: JSON.stringify({
           resetMe: {
             resetMe: true,
+          },
+          grabbableKey: {
+            invertSolidWhileHeld: true
           }
         })
       };
@@ -940,6 +969,9 @@
         userData: JSON.stringify({
           resetMe: {
             resetMe: true,
+          },
+          grabbableKey: {
+            invertSolidWhileHeld: true
           }
         })
       });
@@ -1014,6 +1046,9 @@
         userData: JSON.stringify({
           resetMe: {
             resetMe: true,
+          },
+          grabbableKey: {
+            invertSolidWhileHeld: true
           }
         })
       });
@@ -1053,6 +1088,9 @@
         userData: JSON.stringify({
           resetMe: {
             resetMe: true,
+          },
+          grabbableKey: {
+            invertSolidWhileHeld: true
           }
         })
       });
@@ -1090,6 +1128,9 @@
         userData: JSON.stringify({
           resetMe: {
             resetMe: true,
+          },
+          grabbableKey: {
+            invertSolidWhileHeld: true
           }
         })
       });
@@ -1126,6 +1167,9 @@
         userData: JSON.stringify({
           resetMe: {
             resetMe: true,
+          },
+          grabbableKey: {
+            invertSolidWhileHeld: true
           }
         })
       });
